+++ src/types/material.ts
@@ -6,6 +6,7 @@ export interface Material {
   absorptionCoefficients: number[];
   description: string;
   id: number;
+  origin?: string;
   category: string;
   updatedAt: string;
 }